+++ api/mod.ts
@@ -4,7 +4,7 @@ import { showRoutes } from "hono/dev"
 import { logger } from "hono/logger"
 import { config } from "./config.ts"
 import { openapi_router } from "./routers/openapi/mod.openapi.tsx"
-import { spa_subserver } from "./spa_subserver/spa_subserver.tsx"
+import { page_subserver } from "./spa_subserver/page_subserver.tsx"
 import { serve_static } from "./utils/serve_static.ts"
 import { api_v1 } from "./v1.ts"
 
@@ -18,7 +18,7 @@ app.use(logger()).use(
 
 app.route("/", api_v1)
 app.route("/openapi", openapi_router)
-app.route("/", spa_subserver)
+app.route("/", page_subserver)
 app.notFound(serve_static)
 
 Deno.serve({ port: 3000 }, app.fetch)